Rename folder str_replace_editor to file_editor
- Move openhands/tools/str_replace_editor/ → openhands/tools/file_editor/
- Move tests/tools/str_replace_editor/ → tests/tools/file_editor/
- Update all import statements throughout codebase
- Maintain tool name as 'str_replace_editor' for API compatibility
- All tests pass with new folder structure
