+++ examples/01_hello_world.py
@@ -38,7 +38,7 @@ agent = get_default_agent(
 # from openhands.sdk.tool.registry import register_tool
 # from openhands.sdk.tool.spec import Tool
 # from openhands.tools.execute_bash import BashTool
-# from openhands.tools.str_replace_editor import FileEditorTool
+# from openhands.tools.file_editor import FileEditorTool
 # from openhands.tools.task_tracker import TaskTrackerTool
 # register_tool("BashTool", BashTool)
 # register_tool("FileEditorTool", FileEditorTool)
