+++ examples/02_custom_tools.py
@@ -29,7 +29,7 @@ from openhands.tools.execute_bash import (
     ExecuteBashAction,
     execute_bash_tool,
 )
-from openhands.tools.str_replace_editor import FileEditorTool
+from openhands.tools.file_editor import FileEditorTool
 
 
 logger = get_logger(__name__)
