+++ examples/07_mcp_integration.py
@@ -13,7 +13,7 @@ from openhands.sdk import (
 from openhands.sdk.security.llm_analyzer import LLMSecurityAnalyzer
 from openhands.sdk.tool import Tool, register_tool
 from openhands.tools.execute_bash import BashTool
-from openhands.tools.str_replace_editor import FileEditorTool
+from openhands.tools.file_editor import FileEditorTool
 
 
 logger = get_logger(__name__)
