+++ examples/08_mcp_with_oauth.py
@@ -12,7 +12,7 @@ from openhands.sdk import (
 )
 from openhands.sdk.tool import Tool, register_tool
 from openhands.tools.execute_bash import BashTool
-from openhands.tools.str_replace_editor import FileEditorTool
+from openhands.tools.file_editor import FileEditorTool
 
 
 logger = get_logger(__name__)
