+++ examples/09_pause_example.py
@@ -12,7 +12,7 @@ from openhands.sdk import (
 from openhands.sdk.conversation.state import AgentExecutionStatus
 from openhands.sdk.tool import Tool, register_tool
 from openhands.tools.execute_bash import BashTool
-from openhands.tools.str_replace_editor import FileEditorTool
+from openhands.tools.file_editor import FileEditorTool
 
 
 # Configure LLM
