+++ examples/12_custom_secrets.py
@@ -10,7 +10,7 @@ from openhands.sdk import (
 from openhands.sdk.conversation.secret_source import SecretSource
 from openhands.sdk.tool import Tool, register_tool
 from openhands.tools.execute_bash import BashTool
-from openhands.tools.str_replace_editor import FileEditorTool
+from openhands.tools.file_editor import FileEditorTool
 
 
 # Configure LLM
