+++ examples/14_context_condenser.py
@@ -20,7 +20,7 @@ from openhands.sdk import (
 from openhands.sdk.context.condenser import LLMSummarizingCondenser
 from openhands.sdk.tool import Tool, register_tool
 from openhands.tools.execute_bash import BashTool
-from openhands.tools.str_replace_editor import FileEditorTool
+from openhands.tools.file_editor import FileEditorTool
 from openhands.tools.task_tracker import TaskTrackerTool
