+++ examples/18_send_message_while_processing.py
@@ -52,7 +52,7 @@ from openhands.sdk import (
 )
 from openhands.sdk.tool import Tool, register_tool
 from openhands.tools.execute_bash import BashTool
-from openhands.tools.str_replace_editor import FileEditorTool
+from openhands.tools.file_editor import FileEditorTool
 
 
 # Configure LLM
