+++ openhands/tools/file_editor/editor.py
@@ -9,29 +9,29 @@ from binaryornot.check import is_binary
 
 from openhands.sdk.logger import get_logger
 from openhands.sdk.utils.truncate import maybe_truncate
-from openhands.tools.str_replace_editor.definition import (
+from openhands.tools.file_editor.definition import (
     CommandLiteral,
     FileEditorObservation,
 )
-from openhands.tools.str_replace_editor.exceptions import (
+from openhands.tools.file_editor.exceptions import (
     EditorToolParameterInvalidError,
     EditorToolParameterMissingError,
     FileValidationError,
     ToolError,
 )
-from openhands.tools.str_replace_editor.utils.config import SNIPPET_CONTEXT_WINDOW
-from openhands.tools.str_replace_editor.utils.constants import (
+from openhands.tools.file_editor.utils.config import SNIPPET_CONTEXT_WINDOW
+from openhands.tools.file_editor.utils.constants import (
     BINARY_FILE_CONTENT_TRUNCATED_NOTICE,
     DIRECTORY_CONTENT_TRUNCATED_NOTICE,
     MAX_RESPONSE_LEN_CHAR,
     TEXT_FILE_CONTENT_TRUNCATED_NOTICE,
 )
-from openhands.tools.str_replace_editor.utils.encoding import (
+from openhands.tools.file_editor.utils.encoding import (
     EncodingManager,
     with_encoding,
 )
-from openhands.tools.str_replace_editor.utils.history import FileHistoryManager
-from openhands.tools.str_replace_editor.utils.shell import run_shell_cmd
+from openhands.tools.file_editor.utils.history import FileHistoryManager
+from openhands.tools.file_editor.utils.shell import run_shell_cmd
 
 
 logger = get_logger(__name__)
