+++ openhands/tools/file_editor/utils/file_cache.py
