+++ tests/cross/test_agent_reconciliation.py
@@ -16,8 +16,8 @@ from openhands.sdk.conversation.impl.local_conversation import LocalConversation
 from openhands.sdk.llm import LLM, Message, TextContent
 from openhands.sdk.tool import Tool, register_tool
 from openhands.tools.execute_bash import BashTool
+from openhands.tools.file_editor import FileEditorTool
 from openhands.tools.preset.default import get_default_agent
-from openhands.tools.str_replace_editor import FileEditorTool
 
 
 register_tool("BashTool", BashTool)
