+++ tests/cross/test_hello_world.py
@@ -25,7 +25,7 @@ from openhands.sdk.event.llm_convertible import (
 )
 from openhands.sdk.tool import Tool, register_tool
 from openhands.tools.execute_bash import BashTool
-from openhands.tools.str_replace_editor import FileEditorTool
+from openhands.tools.file_editor import FileEditorTool
 
 
 class TestHelloWorld:
