+++ tests/cross/test_registry_directories.py
@@ -16,7 +16,7 @@ from openhands.sdk.llm import LLM, TextContent
 from openhands.sdk.tool.registry import resolve_tool
 from openhands.sdk.tool.spec import Tool
 from openhands.tools.execute_bash import BashTool
-from openhands.tools.str_replace_editor import FileEditorTool
+from openhands.tools.file_editor import FileEditorTool
 from openhands.tools.task_tracker import TaskTrackerTool
