+++ tests/integration/tests/t03_jupyter_write_file.py
@@ -5,7 +5,7 @@ import os
 from openhands.sdk import get_logger
 from openhands.sdk.tool import Tool, register_tool
 from openhands.tools.execute_bash import BashTool
-from openhands.tools.str_replace_editor import FileEditorTool
+from openhands.tools.file_editor import FileEditorTool
 from tests.integration.base import BaseIntegrationTest, TestResult
