+++ tests/tools/file_editor/test_schema.py
@@ -1,4 +1,4 @@
-from openhands.tools.str_replace_editor import file_editor_tool
+from openhands.tools.file_editor import file_editor_tool
 
 
 def test_to_mcp_tool_detailed_type_validation_editor():
