+++ tests/tools/file_editor/utils/test_encoding.py
@@ -9,8 +9,8 @@ from unittest.mock import patch
 import pytest
 from cachetools import LRUCache
 
-from openhands.tools.str_replace_editor import file_editor
-from openhands.tools.str_replace_editor.utils.encoding import (
+from openhands.tools.file_editor import file_editor
+from openhands.tools.file_editor.utils.encoding import (
     EncodingManager,
     with_encoding,
 )
