+++ tests/tools/file_editor/utils/test_history.py
@@ -3,7 +3,7 @@
 import tempfile
 from pathlib import Path
 
-from openhands.tools.str_replace_editor.utils.history import (
+from openhands.tools.file_editor.utils.history import (
     FileHistoryManager,
 )
