+++ tests/tools/file_editor/utils/test_shell_utils.py
@@ -3,13 +3,13 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from openhands.tools.str_replace_editor.utils.config import (
+from openhands.tools.file_editor.utils.config import (
     MAX_RESPONSE_LEN_CHAR,
 )
-from openhands.tools.str_replace_editor.utils.constants import (
+from openhands.tools.file_editor.utils.constants import (
     CONTENT_TRUNCATED_NOTICE,
 )
-from openhands.tools.str_replace_editor.utils.shell import (
+from openhands.tools.file_editor.utils.shell import (
     check_tool_installed,
     run_shell_cmd,
 )
